+++ facturacion.html
@@ -106,6 +106,7 @@
         </div>
     <script src="bolsaFacturacion.js"></script>
     <script type="module" src="firebase.js"></script>
+    <script type="module" src="insertarFactura.js"></script>
 </body>
 
 </html>
